Remove useModal. Move modal state management logic into Modal component file
